requirements.txt
alembic==1.13.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.3
Bootstrap-Flask==2.3.3
Cartopy==0.22.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorcet==3.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fiona==1.9.5
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
fonttools==4.47.0
fqdn==1.5.1
geographiclib==2.0
geopandas==0.14.2
geopy==2.4.1
geoviews==1.11.0
greenlet==3.0.2
holoviews==1.18.1
hvplot==0.9.1
idna==3.6
infinity==1.5
intervals==0.9.2
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
linkify-it-py==2.0.2
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
movingpandas==0.17.2
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.2
overrides==7.6.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
panel==1.3.6
param==2.0.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.5.0
Pygments==2.17.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
pyviz_comms==3.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Rtree==1.1.0
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.2
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
stack-data==0.6.3
tenacity==8.2.3
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
WTForms==3.1.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
xyzservices==2023.10.1
